Basic operation is working
Constantly in receive mode, when button is pressed then config is switched to transmit mode. When transmission is done (blocking) then receiving is turned back on. Thread control is done using a semaphore. The thread start, stop  and resume api's did not work and cause a config and transmit error since it is apparrently not isr safe.